Move cache to BlockProcessor
